backend/pyproject.toml
[tool.poetry]
name = "library-kiosk-backend"
version = "0.1.0"
description = "Backend API for UAI Library Kiosk System"
authors = ["UAI Development Team <dev@uai.edu>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.6.0"
pydantic = "^2.5.0"
pymysql = "^1.1.0"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
marshmallow = "^3.20.1"
bcrypt = "^4.1.2"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-flask = "^1.3.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]